raw_input: remove IPv6 header when using raw socket
According to RFC 3542, the application should receive the payload
without the IPv6 header and all the extension headers. To receive
IPv6 headers recvmsg interface should be used (not implemented).

This is a temporary workaround!
The IPv6 header should be removed in a different place, to enable
proper implementation of recvmsg for raw socket.

JIRA: COG-42